requirements.txt
Pillow~=10.1.0
pip~=23.3.1
wheel~=0.41.2
setuptools~=68.2.0